requirements.txt
numpy==2.0.2
pyserial==3.5
